introduce `data-color-scheme` selector for global CSS tokens (#196)
Added a new `data-color-scheme` attribute which can be set by consumers on the <html> element or the host element within whose shadow-root the <Root> is rendered. This allows our light/dark color CSS tokens to leak out of <Root> and be applied to the encompassing document or shadow host.

Updated the test-app to use this attribute on <html> and also properly set the color-scheme meta tag, both based on the system preference. Removed the previously existing workarounds in `root.css`.