Use the correct `FontFace` when setting fallbacks
Objects like `FontFace` (or `CSSStyleSheet`) are bound to their own
window, this means for the “popout” or “popup” use case where we’re
rendering to a different window and document using a portal, we need to
ensure we use the `FontFace` from that window. We already are accessing
`ownerDocument.fonts` for setting the fonts this way, but we also need
to use `ownerDocument.defaultView.FontFace` for creating the fonts to
add.

Co-authored-by: Mayank <9084735+mayank99@users.noreply.github.com>