Implement `<Text>` component
I’ve hardcoded the list of typography values in a few places. It might
be helpful to generate types from the typography tokens to avoid some of
this.

At this point, there’s no default variant. It’ll just inherit whatever
is there. If we did want to set a default, it likely would need to be
dynamic based on the current density.