create `<Tree.ItemAction>` component (#355)
Adds a `<Tree.ItemAction>` component which can be passed into the `actions` prop of `<Tree.Item>`. Offers a `visible` prop that can be used to force the visibility of individual actions.

The CSS implementation has changed to use `visibility` on individual actions instead of visually-hidden styles on the actions container. As a result, only visible actions are in tab order.